分页适配example主题
